requirements.txt
telethon==1.26.1
python-binance==1.0.16